refactor api routes to be more consistent.
