+++ apps/gaia/hooks/useGetChild.ts
@@ -2,7 +2,7 @@ import { useQuery } from "react-query"
 
 const useGetChild = (childId: string) => {
   const getChild = async () => {
-    const result = await fetch(`/api/child/${childId}`)
+    const result = await fetch(`/api/children/${childId}`)
     return result.json()
   }
   return useQuery(["child", childId], getChild, {
