+++ apps/gaia/hooks/useGetChildren.ts
@@ -5,7 +5,7 @@ const useGetChildren = () => {
   return useQuery(
     "children",
     async (): Promise<UserData> => {
-      const result = await fetch("/api/children")
+      const result = await fetch("/api/me/children")
 
       if (result.status === 401) {
         const { error } = await result.json()
